apt: Migrated check to new API
