+++ cmk/base/plugins/agent_based/apt.py
@@ -3,12 +3,35 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
+# <<<apt:sep(0)>>>
+# Inst dpkg [1.17.5ubuntu5.3] (1.17.5ubuntu5.4 Ubuntu:14.04/trusty-updates [amd64])
+# Inst libtasn1-6-dev [3.4-3ubuntu0.1] (3.4-3ubuntu0.2 Ubuntu:14.04/trusty-updates [amd64]) []
+# Inst libtasn1-6 [3.4-3ubuntu0.1] (3.4-3ubuntu0.2 Ubuntu:14.04/trusty-updates [amd64])
+# Inst ntpdate [1:4.2.6.p5+dfsg-3ubuntu2.14.04.2] (1:4.2.6.p5+dfsg-3ubuntu2.14.04.3 Ubuntu:14.04/trusty-security [amd64])
+# Inst udev [204-5ubuntu20.10] (204-5ubuntu20.11 Ubuntu:14.04/trusty-updates [amd64]) []
+# Inst libudev1 [204-5ubuntu20.10] (204-5ubuntu20.11 Ubuntu:14.04/trusty-updates [amd64])
+# Inst libpam-systemd [204-5ubuntu20.10] (204-5ubuntu20.11 Ubuntu:14.04/trusty-updates [amd64]) []
+# Inst systemd-services [204-5ubuntu20.10] (204-5ubuntu20.11 Ubuntu:14.04/trusty-updates [amd64]) []
+# Inst libsystemd-daemon0 [204-5ubuntu20.10] (204-5ubuntu20.11 Ubuntu:14.04/trusty-updates [amd64])
+# Inst libsystemd-login0 [204-5ubuntu20.10] (204-5ubuntu20.11 Ubuntu:14.04/trusty-updates [amd64])
+# Inst libpolkit-gobject-1-0 [0.105-4ubuntu2] (0.105-4ubuntu2.14.04.1 Ubuntu:14.04/trusty-updates [amd64])
+# Inst libxext-dev [2:1.3.2-1] (2:1.3.2-1ubuntu0.0.14.04.1 Ubuntu:14.04/trusty-security [amd64]) []
+
+# or
+
+# <<<apt:sep(0)>>>
+# Remv default-java-plugin [2:1.8-58]
+# Remv icedtea-8-plugin [1.6.2-3.1]
+# Inst default-jre [2:1.8-58] (2:1.8-58+deb9u1 Debian:9.11/oldstable [amd64]) []
+# Inst default-jre-headless [2:1.8-58] (2:1.8-58+deb9u1 Debian:9.11/oldstable [amd64])
+
+from collections.abc import Mapping, Sequence
 from dataclasses import dataclass
 from itertools import islice
-from typing import ClassVar, NamedTuple, Optional, Pattern, Sequence
+from typing import Any, ClassVar, NamedTuple, Optional, Pattern
 
-from .agent_based_api.v1 import regex, register
-from .agent_based_api.v1.type_defs import StringTable
+from .agent_based_api.v1 import Metric, regex, register, Result, Service, State
+from .agent_based_api.v1.type_defs import CheckResult, DiscoveryResult, StringTable
 from .utils.apt import ESM_ENABLED, ESM_NOT_ENABLED, NOTHING_PENDING_FOR_INSTALLATION, UBUNTU_PRO
 
 
@@ -170,4 +193,67 @@ def parse_apt(string_table: StringTable) -> Optional[Section]:
 register.agent_section(
     name="apt",
     parse_function=parse_apt,
+)
+
+
+def discover_apt(section: Section) -> DiscoveryResult:
+    yield Service()
+
+
+def _format_summary(action: str, packages: Sequence[str], verbose: bool = False) -> str:
+    summary = f"{len(packages)} {action}"
+    if verbose and packages:
+        summary += " (%s)" % (
+            ", ".join(
+                packages,
+            )
+        )
+
+    return summary
+
+
+def check_apt(params: Mapping[str, Any], section: Section) -> CheckResult:
+    if not any(section):
+        yield Result(state=State.OK, summary=NOTHING_PENDING_FOR_INSTALLATION)
+        return
+
+    if section.no_esm_support:
+        yield Result(
+            state=State.CRIT,
+            summary="System could receive security updates, but needs extended support license",
+        )
+        return
+
+    yield Result(
+        state=State(params["normal"]) if section.updates else State.OK,
+        summary=_format_summary("normal updates", section.updates),
+    )
+    yield Metric(name="normal_updates", value=len(section.updates))
+
+    # Only show removals when necessary as they are very rare.
+    if section.removals:
+        yield Result(
+            state=State(params["removals"]) if section.removals else State.OK,
+            summary=_format_summary("auto removals", section.removals, verbose=True),
+        )
+        yield Metric(name="removals", value=len(section.removals))
+
+    yield Result(
+        state=State(params["security"]) if section.sec_updates else State.OK,
+        summary=_format_summary("security updates", section.sec_updates, verbose=True),
+    )
+    yield Metric(name="security_updates", value=len(section.sec_updates))
+
+
+register.check_plugin(
+    name="apt",
+    service_name="APT Updates",
+    check_function=check_apt,
+    discovery_function=discover_apt,
+    check_default_parameters={
+        "normal": 1,
+        "removals": 1,
+        "security": 2,
+    },
+    check_ruleset_name="apt",
 )
