+++ tests/unit/checks/test_generic_legacy_conversion.py
@@ -150,7 +150,6 @@ def test_all_check_variables_present(fix_plugin_legacy) -> None:
         "METRIC_PULSE_SECURE_DISK",
         "METRIC_PULSE_SECURE_LOG",
         "NAME_TRANSLATION",
-        "NOTHING_PENDING_FOR_INSTALLATION",
         "NO_BLOCKING_SESSIONS_MSG",
         "NimbleReadsType",
         "NimbleWritesType",
@@ -200,7 +199,6 @@ def test_all_check_variables_present(fix_plugin_legacy) -> None:
         "apc_symmetra_ext_temp_default_levels",
         "apc_symmetra_output_default_levels",
         "apc_symmetra_temp_default_levels",
-        "apt_default_levels",
         "arbor_memory_default_levels",
         "arris_cmts_cpu_default_levels",
         "arris_cmts_mem",
@@ -889,7 +887,6 @@ def test_no_new_or_vanished_legacy_checks(fix_plugin_legacy) -> None:
         "appdynamics_memory",
         "appdynamics_sessions",
         "appdynamics_web_container",
-        "apt",
         "arbor_peakflow_sp",
         "arbor_peakflow_sp.disk_usage",
         "arbor_peakflow_sp.flows",
